requirements.txt
ansible==2.9.6
apache-libcloud==2.8.0
apturl==0.5.2
argcomplete==1.8.1
autopep8==1.5
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
django-crispy-forms==1.9.2
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
func-timeout==4.3.5
future==0.18.2
geneticalgorithm==1.0.1
html5lib==1.1
httplib2==0.14.0
idna==2.8
Jinja2==2.10.1
jmespath==0.9.4
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.2
mechanize==0.4.5
monotonic==1.5
netaddr==0.7.19
netifaces==0.10.4
ntlm-auth==1.1.0
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pep8==1.7.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.4
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pywinrm==0.3.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selinux==3.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.2.4
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
xmltodict==0.12.0
